gameserver

+++ server/Engine/Kernal/KernalGuid.cpp
@@ -18,9 +18,9 @@ bool guidInit()
 	return true;
 }
 
-unsigned long int getGuid()
+unsigned int getGuid()
 {
-	unsigned long int curTime = getCurrentTime();
+	unsigned int curTime = getCurrentTime();
 	unsigned int t = curTime;
 	printf("getGuid g_guid=%ld  time=%ld  time1=%d  guid = %ld  t=%d\r\n", g_guid, curTime, curTime, g_guid ^ curTime,t);
 	return g_guid ^ curTime;
