+++ server/Engine/Kernal/KernalUtil.h
@@ -10,5 +10,5 @@ std::vector<std::string> strSplit(const std::string s, const std::string& delim)
 int getLocalIP(char* outip);     // 獲取IP地址
 int getLocalMac(char* mac);      // 獲取mac地址
 
-unsigned long int getCurrentTime();
+unsigned int getCurrentTime();
 #endif
